text was added in readme; make id of objects consistent

+++ src/main/java/com/ing/store_management/repository/StockManagementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StockManagementRepository extends JpaRepository<StockManagement, Integer> {
+public interface StockManagementRepository extends JpaRepository<StockManagement, Long> {
     List<StockManagement> findByProduct(Product product);
 }
